requirements.txt
numpy~=1.23.5
pandas~=1.5.2
torchvision~=0.14.0
torch~=1.13.0
matplotlib~=3.6.2
Pillow~=9.3.0
tqdm~=4.64.1